requirements.txt
# Core Scientific
numpy==1.26.4
scipy>=1.7.3
pandas>=2.2.2
scikit-learn>=1.1.1

# Audio: Feature, ASR, Diarization
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0
openai-whisper>=20231117

# Diarization (pyannote) -- versi stabil!
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
einops>=0.6.0
asteroid-filterbanks>=0.3.0

# NLP/Transformers/Fusion
transformers>=4.30.0
sentencepiece>=0.1.99
datasets>=4.2.0,<5.0.0
accelerate>=0.20.0

# Deep Learning
torch>=2.0.0
lightning>=2.0.0
pytorch-lightning>=2.0.0

# Scraping & Data Collect
yt-dlp>=2023.7.6
requests>=2.32.0
beautifulsoup4>=4.12.3
readability-lxml>=0.8.1
trafilatura>=1.12.0

# Progress & Utilities
tqdm>=4.66.4
python-dotenv>=1.0.1
pytz>=2022.1,<2023.0
pyyaml>=5.1
packaging>=20.0

# Visualization & Exploration
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter & Kernel (for Notebook)
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: For compatibility
# (No duplicate lines; version conflict resolved; only include one pandas, tqdm, yt-dlp, etc)
